+++ laktory/models/sql/__init__.py
@@ -6,5 +6,6 @@ from .tableaggregation import TableAggregation
 from .tablebuilder import TableBuilder
 from .tableexpectation import TableExpectation
 from .tablejoin import TableJoin
+from .tableunion import TableUnion
 from .tablewindowfilter import TableWindowFilter
 from .volume import Volume