+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.1.9] - Unreleased
 ### Added
+* Support for table unions in table builder
 * New column property `raise_missing_arg_exception` to allow for some spark function inputs to be missing
 
 ## [0.1.8] - 2024-03-25